increase size to accomodate bigger strings

+++ lps.h
@@ -72,7 +72,7 @@ struct lps {
  * @param str The input string to be parsed.
  * @param len The length of the string to be parsed.
  */
-void init_lps(struct lps *lps_ptr, const char *str, int len);
+void init_lps(struct lps *lps_ptr, const char *str, uint64_t len);
 
 /**
  * @brief Constructs an lps object from a string.
@@ -82,7 +82,7 @@ void init_lps(struct lps *lps_ptr, const char *str, int len);
  * @param len The length of the string to be parsed.
  * @param offset The length of the offset in which each index will be shifted.
  */
-void init_lps_offset(struct lps *lps_ptr, const char *str, int len, uint64_t offset);
+void init_lps_offset(struct lps *lps_ptr, const char *str, uint64_t len, uint64_t offset);
 
 /**
  * @brief Constructs an lps object from a string, with reverse complement
@@ -92,7 +92,7 @@ void init_lps_offset(struct lps *lps_ptr, const char *str, int len, uint64_t off
  * @param str The input string to be parsed.
  * @param len The length of the string to be parsed.
  */
-void init_lps2(struct lps *lps_ptr, const char *str, int len);
+void init_lps2(struct lps *lps_ptr, const char *str, uint64_t len);
 /**
  * @brief Initializes an lps object by reading its contents from a binary file.
  *
@@ -120,7 +120,7 @@ void init_lps3(struct lps *lps_ptr, FILE *in);
  * @param len The length of the string to be parsed.
  * @param chunk_size The length of the chunks to be processed.
  */
-void init_lps4(struct lps *lps_ptr, const char *str, int len, int lcp_level, int chunk_size);
+void init_lps4(struct lps *lps_ptr, const char *str, uint64_t len, int lcp_level, int chunk_size);
 
 /**
  * @brief Destructor for the lps object. Frees dynamically allocated memory for cores.
